requirements.txt
pytesseract == 0.3.10
pyPDF2 == 3.0.1
psycopg2 == 2.9.9
pillow == 10.4.0
pdfkit == 1.0.0
pdf2image == 1.17.0
pdf2img == 0.1.2
Django == 4.2.3
djangorestframework == 3.14.0
djangorestframework-simplejwt == 5.3.0
django-cors-headers == 4.2.0
azure-ai-formrecognizer == 3.3.3
azure-common == 1.1.28
azure-core == 1.31.0
pandas == 2.2.2
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
PyMuPDF==1.23.22
python-docx==1.1.0
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0

